refactor: move temporal attribution to SAST owner
